tests(generators): revert change to plugin helper and move tests to new file
