+++ packages/generators/src/utils/plugins.ts
@@ -40,7 +40,9 @@ export const replaceAt = (str: string, index: number, replace: string): string =
 export const generatePluginName = (rawPluginName: string): string => {
     const myPluginNameArray = rawPluginName.split('-');
     const pluginArrLength: number = myPluginNameArray.length;
-    for (let i = 0; i < pluginArrLength; i++) {
+    // ignore plugin names without hyphens to allow for cases
+    // such as webpack.DefinePlugin, which should not be capitalized
+    for (let i = 0; i < pluginArrLength && pluginArrLength > 1; i++) {
         myPluginNameArray[i] = replaceAt(myPluginNameArray[i], 0, myPluginNameArray[i].charAt(0).toUpperCase());
     }
     return myPluginNameArray.join('');
